+++ src/hooks/useCalendar.ts
@@ -1,7 +1,7 @@
 import { useCallback, useState } from 'react';
-import { fetchZhawSchedule, fetchZhawStudents, fetchZhawLecturers } from '@/lib/handlers/calendarHandler';
+import { fetchZhawStudents, fetchZhawLecturers } from '@/lib/handlers/calendarHandler';
 import { fetchPublicHolidays } from '@/lib/api/openholidays';
-import { mapZhawDaysToEvents } from '@/lib/calendar';
+import { generateCalendarEvents } from '@/lib/calendar';
 import type { EventInput, EventSourceFuncArg } from '@fullcalendar/core';
 import { useTranslation } from 'react-i18next';
 
@@ -43,7 +43,7 @@ export function useCalendar(shortName: string) {
 
       const [publicHolidays, scheduleEvents] = await Promise.all([
         fetchPublicHolidays(viewStart.getFullYear()),
-        rolePath ? fetchScheduleEvents(shortName, viewStart, viewEnd, rolePath, t) : Promise.resolve([]),
+        rolePath ? generateCalendarEvents(shortName, viewStart, viewEnd, rolePath, t) : Promise.resolve([]),
       ]);
 
       successCallback([...publicHolidays, ...scheduleEvents]);
@@ -59,49 +59,3 @@ export function useCalendar(shortName: string) {
     fetchEventsDynamically,
   };
 }
-
-const fetchScheduleEvents = async (
-  shortName: string,
-  viewStart: Date,
-  viewEnd: Date,
-  rolePath: string,
-  t: (key: string) => string
-) => {
-  const allEvents: EventInput[] = [];
-  const fetchedWeeks = new Set<string>();
-  const maxDays = 7;
-
-  for (let date = new Date(viewStart); date <= viewEnd; date.setDate(date.getDate() + maxDays)) {
-    const adjustedDate = new Date(date);
-    adjustedDate.setDate(adjustedDate.getDate() + 1);
-    const startingAt = adjustedDate.toISOString().split('T')[0];
-
-    if (fetchedWeeks.has(startingAt)) {continue;}
-    fetchedWeeks.add(startingAt);
-
-    const data = await fetchZhawSchedule(shortName, startingAt, rolePath);
-    const isWeekEmpty =
-      !data.days || data.days.length === 0 ||
-      data.days.every((day) => !day.events || day.events.every((e) => e.type === 'Holiday'));
-
-    if (isWeekEmpty) {
-      const end = new Date(adjustedDate);
-      end.setDate(end.getDate() + 7);
-      allEvents.push({
-        title: t('semesterBreak'),
-        start: startingAt,
-        end: end.toISOString().split('T')[0],
-        allDay: true,
-        color: '#F85A6D',
-      });
-    } else {
-      const filtered = data.days.map((day) => ({
-        ...day,
-        events: (day.events ?? []).filter((e) => e.type !== 'Holiday'),
-      }));
-      allEvents.push(...mapZhawDaysToEvents({ days: filtered }));
-    }
-  }
-
-  return allEvents;
-};
